Начаоло выполнения Homework-1 / task-1
