+++ homework-1/task-1/main.go
@@ -1,32 +1,37 @@
 package main
 
 import (
+	"net/http"
+	// "io"
 	"fmt"
+	// "os"
+	"log"
 )
 
+// Напишите функцию, которая будет получать на вход строку с поисковым запросом (string) и массив ссылок на страницы, по которым стоит
+// произвести поиск ([]string). Результатом работы функции должен быть массив строк со ссылками на страницы, на которых обнаружен 
+// поисковый запрос. Функция должна искать точное соответствие фразе в тексте ответа от сервера по каждой из ссылок. 
+
+func testSearch(searchQuery string, whereSearch []string) {
+	for	_, url := range whereSearch {
+		openURL(url)
+	}
+}
+
+func openURL(url) {
+	get Url
+}
+
+
 func main() {
-	a := GetRequest()
-	b := GetReferenceSite()
-	fmt.Println(a)
-	fmt.Println(b)
-}
 
-//GetRequest - получает строку поискового запроса
-func GetRequest() string {
-	var StrRequest string
-	fmt.Println("Введите поисковый запрос:")
-	fmt.Scan(&StrRequest)
-	return StrRequest
-}
+	resp, err := http.Get("https://golang.org/")
+	if err != nil {
+		log.Print(err)
+		return
+	}
+	defer resp.Body.Close()
 
-//GetReferenceSite - получает сайты на которых необходимо искать
-func GetReferenceSite() []string {
-	var ArrReference []string
-	var GetSingleRef string
-	fmt.Println("введите названия сайтов на которых поискать")
-	fmt.Scanln(&GetSingleRef)
-	ArrReference = append(ArrReference, GetSingleRef)
-	// 		for i := 0; i > 3; i++ {}
-	return ArrReference
-
+	fmt.Println(resp.Status)
+	
 }
